The const_engine lead to a number of programming errors. Renamed to epsilon_engine because of this.
The typical error is that epsilon_engines can be aborted and that there
are no cleanup actions for that case. Fixed that for almost all cases.

direct_socket_acceptor is now atomic, i.e. it either accepts or does
not accept, there is no longer something in between. Netplex assumes
this, so this is a fix for a netplex problem.

Freshly created Rpc servers are now ensured to do I/O first when the 
event loop is again entered.



git-svn-id: https://godirepo.camlcity.org/svn/lib-ocamlnet2/trunk/code@973 96e5b9e0-aa12-0410-9e2e-b23d24e56262