Fixing workload management. Workload managers have now a method [capacity] that reports the capacity of a certain container. This way, these managers can directly influence scheduling.
The [adjust] method is now called whenever the load parameters
of a container change. That makes it impossible that scheduling
is based on wrong load parameters.

Dynamic workload manager: max_jobs_per_thread is now interpreted
as absolute maximum that must not be exceeded. There is a second
parameter recommended_jobs_per_thread that can be exceeded in
overload situations.

Dynamic workload manager: implementing a timeout for inactivation.



git-svn-id: https://godirepo.camlcity.org/svn/lib-ocamlnet2/trunk/code@974 96e5b9e0-aa12-0410-9e2e-b23d24e56262